requirements.txt

dj-database-url==0.5.0
Django
# Django==4.0.3

django-environ==0.8.1
django-filter==21.1
django-gridfs-storage==0.0.2
django-heroku==0.0.0
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
django-url-filter==0.3.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo
# djongo==1.3.6

drf-yasg2==1.19.4
enum-compat==0.0.3
gunicorn==20.1.0
Pillow==8.4.0
pymongo==3.12.3
google-api-python-client==2.38.0
alpaca-trade-api==1.5.1
requests==2.26.0
mongoengine==0.23.1

# Important for Heroky
# psycopg2-binary
# Important for Dockerfile. 


dnspython==2.1.0

# Trading data libraries:
yahoo-fin==0.8.9.1

# Library for testing 
coverage
apscheduler

# TRY 
# pyparsing==3.0.6
# python-dateutil==2.8.2
# pytz==2021.3
# PyYAML==6.0
# ruamel.yaml==0.17.19
# ruamel.yaml.clib==0.2.6
# simplejson==3.17.6
# six==1.16.0
# sqlparse==0.2.4
# toml==0.10.2
# uritemplate==4.1.1
# urllib3==1.26.7
# websocket-client==1.2.3
# websockets==9.1
# whitenoise==5.3.0
# psycopg2==2.9.3
# httpie
# pycodestyle==2.8.0
# PyJWT==2.3.0
# idna==3.3
# inflection==0.5.1
# itypes==1.2.0
# Jinja2==3.0.3
# MarkupSafe==2.0.1
# msgpack==1.0.2
# numpy==1.22.2
# openapi-codec==1.3.2
# packaging==21.3
# pandas==1.4.1
# asgiref==3.4.1
# autopep8==1.6.0
# cached-property==1.5.2
# certifi==2021.10.8
# charset-normalizer==2.0.9
# coreapi==2.3.3
# coreschema==0.0.4